Moved GetEnv to StdLib
